+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [2.3.1] - xxxx-xx-xx
+
+### Fixed
+
+- A small bug when comparing two backbones with the same set of negative/positive/optional variables but different satisfiability.
+
 LogicNG uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [2.3.0] - 2022-07-18